feat: debug prompt
